chatbot/requirements.txt
streamlit==1.33.0
requests==2.31.0
python-dotenv==1.0.0 